backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
spacy>=3.6.0
scikit-learn>=1.3.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
pydantic>=2.0.0
